Start trying to make the fake streaming machinery not-logs-specific
